1. 正式发布ln882h wifi mac stack，完全解除与platform（soc、os、net...）的耦合，全面深度重构和优化，精简api。 2. 全新的cmake构建体系 。 3. 更新keil MDK(win平台)，gcc(win/linux平台)开发环境指导文档。 4. 增加LN882H分区地址辅助计算工具。 5. 全新的编译预处理、后处理脚本工具体系。 6. 新版AT命令组件。 7. 修复armcc5，gcc编译警告若干。 8. 更新peripher hal driver. 9. 更新Jflash烧录配置参数。

+++ components/net/dhcpd/dhcp_packet.c
@@ -105,7 +105,7 @@ int dhcpd_parse_option(uint8_t *packet, uint16_t packet_len, dhcp_option_t *opti
     return LN_TRUE;
 }
 
-int dhcpd_recv_packet_check(uint8_t *packet, int len)
+int dhcpd_recv_packet_check(uint8_t *packet, uint16_t len)
 {
     int ret = LN_FALSE;
     if (len < (BOOTP_ABSOLUTE_MIN_LEN + DHCP_MAGIC_COOKIE_LEN + 4)) {
